+++ test_tf2/test/buffer_core_test.cpp
@@ -48,7 +48,7 @@
 #include <tf2/exceptions.h>
 #include <tf2/LinearMath/Quaternion.h>
 #include <tf2/time.h>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #include <tf2_ros/buffer_interface.h>
 
 #include "permuter.hpp"
