+++ test_tf2/test/test_buffer_client.cpp
@@ -38,7 +38,7 @@
 
 #include <rclcpp/rclcpp.hpp>
 #include <tf2_bullet/tf2_bullet.hpp>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 #include <tf2_kdl/tf2_kdl.hpp>
 #include <tf2_ros/buffer_client.h>
