+++ tf2_geometry_msgs/test/test_tf2_geometry_msgs.cpp
@@ -42,7 +42,7 @@
 #include <tf2_ros/transform_listener.h>
 #include <gtest/gtest.h>
 #include <tf2_ros/buffer.h>
-#include <tf2_geometry_msgs/tf2_geometry_msgs.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
 
 std::unique_ptr<tf2_ros::Buffer> tf_buffer = nullptr;
 static const double EPS = 1e-3;
